idle add output and waybar trigger

+++ .scripts/toggle_idle.sh
@@ -1,15 +1,41 @@
-
-
 CMD="hypridle"
+
+OUTPUT_IDLE_RUN_STATE=false
+
+while getopts "o" flag; do
+    case "${flag}" in
+        o) OUTPUT_IDLE_RUN_STATE=true ;;
+    esac
+done
 
 is_not_running() {
     PROCESS_ID=$(pgrep "${CMD}")
     [ -z "$PROCESS_ID" ]
 }
 
-if is_not_running; then
-    "${CMD}" &>/dev/null & disown;
-else
-    pkill "${CMD}"
-fi
 
+waybar_sig(){
+    pkill -RTMIN+3 waybar
+}
+
+main() {
+    if [ $OUTPUT_IDLE_RUN_STATE = true ]; then
+        if is_not_running; then
+            echo "0"
+        else
+            echo "1"
+        fi
+        exit
+    fi
+
+    if is_not_running; then
+        "${CMD}" &>/dev/null & disown;
+    else
+        pkill "${CMD}"
+    fi
+
+    waybar_sig
+}
+
+
+main
